requirements.txt
# Customer Analytics & Recommendation System Requirements

# Core Data Science Libraries
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Visualization Libraries
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Configuration Management
PyYAML>=6.0

# Data Generation (optional for development)
faker>=18.0.0

# Testing (optional for development)
pytest>=7.0.0
pytest-cov>=4.0.0

# Type Checking (optional for development)
mypy>=1.0.0

# Code Formatting (optional for development)
black>=22.0.0

# Linting (optional for development)
flake8>=4.0.0 